Move histogram to separate file
